svg/interact/parsing/pointer-events-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>SVG Scripting and Interactivity: parsing pointer-events with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/interact.html#PointerEventsProperty" rel="help"/><h:meta content="pointer-events supports only the grammar 'auto | bounding-box | visiblePainted | visibleFill | visibleStroke | visible | painted | fill | stroke | all | none'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("pointer-events", "markers");
test_invalid_value("pointer-events", "0");
test_invalid_value("pointer-events", "visiblePainted visibleFill");

  ]]></script></svg>
